+++ src/components/school/repo/school.repo.js
@@ -1,21 +1,23 @@
 const { schoolModel: School } = require('../../../services/school/model');
+
+const { SCHOOL_OF_DELETE_USERS } = require('./db');
 const { NotFound } = require('../../../errors');
 
 const getSchool = async (_id) => {
-	const school = await School.findOne({ _id }).lean().exec();
+	const school = await School.findById(_id).lean().exec();
 	if (school == null) {
 		throw new NotFound('no school for this id');
 	}
 	return school;
 };
 
-const findSchools = async (selectCriteria) => School.find(selectCriteria).lean().exec();
+const getTombstoneSchool = async () => School.findOne({ purpose: SCHOOL_OF_DELETE_USERS.purpose }).lean().exec();
 
-const updateSchool = async (schoolId, schoolPatch) =>
-	School.findByIdAndUpdate(schoolId, schoolPatch, { new: true }).lean().exec();
+const setTombstoneUser = async (schoolId, tombstoneUserId) =>
+	School.findByIdAndUpdate(schoolId, { tombstoneUserId }, { new: true }).lean().exec();
 
 module.exports = {
 	getSchool,
-	findSchools,
-	updateSchool,
+	getTombstoneSchool,
+	setTombstoneUser,
 };